requirements.txt
PyMuPDF==1.21.1
Pillow==9.1.1
requests==2.31.0
python-dotenv==1.0.0